Refactor DataNavigation component
Moved it outside of the Header component and moved api calls to within itself
