+++ src/components/data/Header.js
@@ -38,8 +38,6 @@ function Header(props) {
     const classes = useStyles();
     const {
         baseId,
-        path,
-        error,
         isGridView,
         toggleDisplay,
         onDownloadSelected,
@@ -49,102 +47,83 @@ function Header(props) {
         detailsEnabled,
         onDetailsSelected,
         intl,
-        dataRoots,
-        userHomePath,
-        userTrashPath,
-        sharedWithMePath,
-        communityDataPath,
     } = props;
 
     let headerId = build(baseId, ids.HEADER);
 
     return (
-        <>
-            <Toolbar variant="dense">
-                <DataNavigation
-                    path={path}
-                    error={error}
-                    baseId={baseId}
-                    dataRoots={dataRoots}
-                    userHomePath={userHomePath}
-                    userTrashPath={userTrashPath}
-                    sharedWithMePath={sharedWithMePath}
-                    communityDataPath={communityDataPath}
-                />
-            </Toolbar>
-            <Toolbar
-                variant="dense"
-                classes={{ root: classes.toolbar }}
-                id={headerId}
-            >
-                {isGridView && (
-                    <Tooltip
-                        id={build(headerId, ids.TABLE_VIEW_BTN, ids.TOOLTIP)}
-                        title={getMessage("tableView")}
-                        aria-label={formatMessage(intl, "tableView")}
+        <Toolbar
+            variant="dense"
+            classes={{ root: classes.toolbar }}
+            id={headerId}
+        >
+            {isGridView && (
+                <Tooltip
+                    id={build(headerId, ids.TABLE_VIEW_BTN, ids.TOOLTIP)}
+                    title={getMessage("tableView")}
+                    aria-label={formatMessage(intl, "tableView")}
+                >
+                    <IconButton
+                        id={build(headerId, ids.TABLE_VIEW_BTN)}
+                        edge="start"
+                        className={classes.menuButton}
+                        onClick={() => toggleDisplay()}
                     >
-                        <IconButton
-                            id={build(headerId, ids.TABLE_VIEW_BTN)}
-                            edge="start"
-                            className={classes.menuButton}
-                            onClick={() => toggleDisplay()}
-                        >
-                            <TableIcon />
-                        </IconButton>
-                    </Tooltip>
-                )}
-                {!isGridView && (
-                    <Tooltip
-                        id={build(headerId, ids.GRID_VIEW_BTN, ids.TOOLTIP)}
-                        title={getMessage("gridView")}
-                        aria-label={formatMessage(intl, "gridView")}
+                        <TableIcon />
+                    </IconButton>
+                </Tooltip>
+            )}
+            {!isGridView && (
+                <Tooltip
+                    id={build(headerId, ids.GRID_VIEW_BTN, ids.TOOLTIP)}
+                    title={getMessage("gridView")}
+                    aria-label={formatMessage(intl, "gridView")}
+                >
+                    <IconButton
+                        id={build(headerId, ids.GRID_VIEW_BTN)}
+                        edge="start"
+                        className={classes.menuButton}
+                        onClick={() => toggleDisplay()}
                     >
-                        <IconButton
-                            id={build(headerId, ids.GRID_VIEW_BTN)}
-                            edge="start"
-                            className={classes.menuButton}
-                            onClick={() => toggleDisplay()}
-                        >
-                            <GridIcon />
-                        </IconButton>
-                    </Tooltip>
-                )}
-                <div className={classes.divider} />
-                {detailsEnabled && (
-                    <Button
-                        id={build(headerId, ids.DETAILS_BTN)}
-                        variant="contained"
-                        disableElevation
-                        color="primary"
-                        className={classes.button}
-                        onClick={onDetailsSelected}
-                    >
-                        <Info className={classes.buttonIcon} />
-                        <Hidden xsDown>{getMessage("details")}</Hidden>
-                    </Button>
-                )}
-                <UploadMenuBtn baseId={headerId} />
+                        <GridIcon />
+                    </IconButton>
+                </Tooltip>
+            )}
+            <div className={classes.divider} />
+            {detailsEnabled && (
                 <Button
-                    id={build(headerId, ids.SHARE_BTN)}
+                    id={build(headerId, ids.DETAILS_BTN)}
                     variant="contained"
                     disableElevation
                     color="primary"
                     className={classes.button}
-                    onClick={() => console.log("Share")}
+                    onClick={onDetailsSelected}
                 >
-                    <Share className={classes.buttonIcon} />
-                    <Hidden xsDown>{getMessage("share")}</Hidden>
+                    <Info className={classes.buttonIcon} />
+                    <Hidden xsDown>{getMessage("details")}</Hidden>
                 </Button>
-                <DataDotMenu
-                    baseId={headerId}
-                    onDownloadSelected={onDownloadSelected}
-                    onEditSelected={onEditSelected}
-                    onMetadataSelected={onMetadataSelected}
-                    onDeleteSelected={onDeleteSelected}
-                    ButtonProps={{ className: classes.whiteDotMenu }}
-                />
-            </Toolbar>
-        </>
+            )}
+            <UploadMenuBtn baseId={headerId} />
+            <Button
+                id={build(headerId, ids.SHARE_BTN)}
+                variant="contained"
+                disableElevation
+                color="primary"
+                className={classes.button}
+                onClick={() => console.log("Share")}
+            >
+                <Share className={classes.buttonIcon} />
+                <Hidden xsDown>{getMessage("share")}</Hidden>
+            </Button>
+            <DataDotMenu
+                baseId={headerId}
+                onDownloadSelected={onDownloadSelected}
+                onEditSelected={onEditSelected}
+                onMetadataSelected={onMetadataSelected}
+                onDeleteSelected={onDeleteSelected}
+                ButtonProps={{ className: classes.whiteDotMenu }}
+            />
+        </Toolbar>
     );
 }
