+++ src/components/data/listing/Listing.js
@@ -6,13 +6,8 @@
  */
 import React, { useEffect, useState } from "react";
 
-import { formatMessage, withI18N } from "@cyverse-de/ui-lib";
-import { TablePagination } from "@material-ui/core";
-import HomeIcon from "@material-ui/icons/Home";
-import FolderSharedIcon from "@material-ui/icons/FolderShared";
-import GroupIcon from "@material-ui/icons/Group";
-import DeleteIcon from "@material-ui/icons/Delete";
-import { useRouter } from "next/router";
+import { withI18N } from "@cyverse-de/ui-lib";
+import { TablePagination, Toolbar } from "@material-ui/core";
 import { injectIntl } from "react-intl";
 
 import Header from "../Header";
@@ -21,20 +16,15 @@ import TableView from "./TableView";
 import UploadDropTarget from "../../uploads/UploadDropTarget";
 import { useUploadTrackingState } from "../../../contexts/uploadTracking";
 import { camelcaseit } from "../../../common/functions";
-import NavigationConstants from "../../../common/NavigationConstants";
-import { useUserProfile } from "../../../contexts/userProfile";
-import constants from "../../../constants";
 import Drawer from "../details/Drawer";
 import {
-    getFilesystemRoots,
     getInfoTypes,
     getPagedListing,
 } from "../../endpoints/Filesystem";
+import DataNavigation from "../DataNavigation";
 
 function Listing(props) {
-    const router = useRouter();
     const uploadTracker = useUploadTrackingState();
-    const [userProfile] = useUserProfile();
 
     const [isGridView, setGridView] = useState(false);
     const [order, setOrder] = useState("asc");
@@ -51,12 +41,7 @@ function Listing(props) {
     const [detailsResource, setDetailsResource] = useState(null);
     const [infoTypes, setInfoTypes] = useState([]);
 
-    const [dataRoots, setDataRoots] = useState([]);
-    const [userHomePath, setUserHomePath] = useState("");
-    const [userTrashPath, setUserTrashPath] = useState("");
-    const [sharedWithMePath, setSharedWithMePath] = useState("");
-    const [communityDataPath, setCommunityDataPath] = useState("");
-    const { baseId, path, handlePathChange, intl } = props;
+    const { baseId, path, handlePathChange } = props;
 
     // Used to force the data listing to refresh when uploads are completed.
     const uploadsCompleted = uploadTracker.uploads.filter((upload) => {
@@ -96,38 +81,6 @@ function Listing(props) {
             );
         }
     }, [path, rowsPerPage, orderBy, order, page, uploadsCompleted]);
-
-    useEffect(() => {
-        getFilesystemRoots().then((respData) => {
-            if (respData && userProfile) {
-                const respRoots = respData.roots;
-                const home = respRoots.find(
-                    (root) => root.label === userProfile.id
-                );
-                home.icon = <HomeIcon />;
-                const sharedWithMe = respRoots.find(
-                    (root) => root.label === constants.SHARED_WITH_ME
-                );
-                setSharedWithMePath(sharedWithMe.path);
-                sharedWithMe.icon = <FolderSharedIcon />;
-                const communityData = respRoots.find(
-                    (root) => root.label === constants.COMMUNITY_DATA
-                );
-                setCommunityDataPath(communityData.path);
-                communityData.icon = <GroupIcon />;
-                const trash = respRoots.find(
-                    (root) => root.label === formatMessage(intl, "trash")
-                );
-                trash.icon = <DeleteIcon />;
-
-                const basePaths = respData["base-paths"];
-                setUserHomePath(basePaths["user_home_path"]);
-                setUserTrashPath(basePaths["user_trash_path"]);
-                setDataRoots([home, sharedWithMe, communityData, trash]);
-            }
-            setError("");
-        });
-    }, [path, intl, userProfile]);
 
     useEffect(() => {
         getInfoTypes().then((resp) => {
@@ -240,15 +193,6 @@ function Listing(props) {
         setPage(0);
     };
 
-    //route to default path
-    if (dataRoots.length > 0 && (!error || error.length === 0) && !path) {
-        router.push(
-            encodeURI(
-                `${constants.PATH_SEPARATOR}${NavigationConstants.DATA}${constants.PATH_SEPARATOR}ds${dataRoots[0].path}`
-            )
-        );
-    }
-
     const onDetailsSelected = () => {
         setDetailsOpen(true);
         const selectedId = selected[0];
@@ -262,21 +206,20 @@ function Listing(props) {
     return (
         <>
             <UploadDropTarget path={path}>
+                <Toolbar variant="dense">
+                    <DataNavigation
+                        path={path}
+                        baseId={baseId}
+                    />
+                </Toolbar>
                 <Header
                     baseId={baseId}
                     isGridView={isGridView}
                     toggleDisplay={toggleDisplay}
-                    path={path}
-                    error={error}
                     onDownloadSelected={onDownloadSelected}
                     onEditSelected={onEditSelected}
                     onMetadataSelected={onMetadataSelected}
                     onDeleteSelected={onDeleteSelected}
-                    dataRoots={dataRoots}
-                    userHomePath={userHomePath}
-                    userTrashPath={userTrashPath}
-                    sharedWithMePath={sharedWithMePath}
-                    communityDataPath={communityDataPath}
                     detailsEnabled={detailsEnabled}
                     onDetailsSelected={onDetailsSelected}
                 />
