+++ src/constants.js
@@ -5,5 +5,6 @@ export default {
     CYVERSE_URL: "https://cyverse.org/discovery-environment",
     SHARED_WITH_ME: "Shared With Me",
     COMMUNITY_DATA: "Community Data",
+    TRASH: "Trash",
     PATH_SEPARATOR: "/",
 };
